+++ util/src/org/commcare/util/screen/EntityListSubscreen.java
@@ -1,5 +1,6 @@
 package org.commcare.util.screen;
 
+import org.commcare.modern.util.Pair;
 import org.commcare.suite.model.Action;
 import org.commcare.suite.model.Detail;
 import org.commcare.suite.model.DetailField;
@@ -80,6 +81,35 @@ public class EntityListSubscreen extends Subscreen<EntityScreen> {
             }
         }
         return row.toString();
+    }
+
+    public static Pair<String[], int[]> getHeaders(Detail shortDetail, EvaluationContext context){
+        DetailField[] fields = shortDetail.getFields();
+        String[] headers = new String[fields.length];
+        int[] widthHints = new int[fields.length];
+
+        StringBuilder row = new StringBuilder();
+        int i = 0;
+        for (DetailField field : fields) {
+            String s = field.getHeader().evaluate(context);
+
+            int widthHint = SCREEN_WIDTH / fields.length;
+            try {
+                widthHint = Integer.parseInt(field.getHeaderWidthHint());
+            } catch (Exception e) {
+                //Really don't care if it didn't work
+            }
+            ScreenUtils.addPaddedStringToBuilder(row, s, widthHint);
+
+            headers[i] = s;
+            widthHints[i] = widthHint;
+
+            i++;
+            if (i != fields.length) {
+                row.append(" | ");
+            }
+        }
+        return new Pair<>(headers, widthHints);
     }
 
     //So annoying how identical this is...
